backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Sarah Hackett <sarah.hackett@flutteruki.com>"]
readme = "README.md"
packages = [{ include = "backend" }] 

[tool.poetry.dependencies]
python = ">=3.12,<4.0"
flask = ">=3.1.1,<4.0.0"
flask-cors = ">=6.0.1,<7.0.0"
requests = ">=2.32.4,<3.0.0"
python-dotenv = ">=1.1.1,<2.0.0"
marshmallow = "^4.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-cov = "^6.2.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q --cov=app --cov-report=term-missing"
